Move into core module
Creating a module structure for project, so the first order of
business is to create the core module. Other modules to come.
